refactor(match): improve readability and error handling
- Extract common elements of getMatchableSenders and -Receivers
- Rename and document some helper functions
- Add super to generate endpoint
- Allow nullability of Match.meetingInfo.date in mongoose schema
- Be stricter when verifying MatchGenerateOperation input data
- Use the correct error code for bad request
