+++ src/collections/match/match.schema.ts
@@ -30,7 +30,7 @@ const matchBaseSchema = {
   meetingInfo: {
     location: { type: String, required: true },
     // Can be null in case of a StandMatch
-    date: { type: Date, required: true },
+    date: Date,
   },
 };
